Extract Path parsing into new class
Extract the logic to parse a Path out of the Parser and into
a new PathParser class.
