+++ config/src/test/scala/com/typesafe/config/impl/PathTest.scala
@@ -73,8 +73,8 @@ class PathTest extends TestUtils {
 
         for (t <- tests) {
             assertEquals(t.expected, t.path.render())
-            assertEquals(t.path, Parser.parsePath(t.expected))
-            assertEquals(t.path, Parser.parsePath(t.path.render()))
+            assertEquals(t.path, PathParser.parsePath(t.expected))
+            assertEquals(t.path, PathParser.parsePath(t.path.render()))
         }
     }
